requirements.txt
anthropic>=0.40.0
reportlab>=4.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
pdfplumber>=0.10.0
feedparser>=6.0.0
